Check for available updates, updated files
The program will now check for updates when it starts up. If any are
available, a messagebox will pop up and notify the user. The license has
also been updated to include the name of the program and the author. The
readme has been finished up, and a version file has been added. Version
number incremented.